requirements.txt
djangorestframework
pytz
sqlparser
Django
asgiref